requirements.txt
Django==3.2.9
django-user-agents==0.4.0
pytz==2021.3
PyYAML==6.0
ua-parser==0.8.0
user-agents==1.1.0
pytest-django==4.4.0
django-extensions==3.1.5
djangorestframework==3.12.4
